+++ audio/sample_rate_receiver.go
@@ -28,8 +28,8 @@ type sampleRateReceiver struct {
 func (p *sampleRateReceiver) ReceivePCMFrame(userID snowflake.ID, packet *PCMPacket) error {
 	newPCM := make([]int16, opus.GetOutputBuffSize(p.outputSampleRate, p.resampler.Channels()))
 	var (
-		inputFrames  int
-		outputFrames int
+		inputFrames  int64
+		outputFrames int64
 	)
 	if err := p.resampler.Process(packet.PCM, newPCM, p.inputSampleRate, p.outputSampleRate, 0, &inputFrames, &outputFrames); err != nil {
 		return err
